Expand on Euler007

+++ solutions/euler007/solution.go
@@ -23,7 +23,7 @@ func main() {
 			os.Exit(2)
 		}
 
-		prime := ordinator.Ordinal(ordinal)
+		prime := ordinator.Ordinal(uint(ordinal))
 		fmt.Println(prime)
 	}
 }
